requirements.txt
Flask
flask
requests
markdown
slither-analyzer
solc-select
pyecharts
python-dotenv